WIP: clang-tidy 잠재 14 fix + audit phase 문서화
audit 2nd/3rd 결과 반영:
- audit 2nd (90s smoke): cppcheck 79 / clang-tidy 166 / ASan 5 direct + 174 indirect leak (1.2 MB) - 모두 external lib + NPU buffer (90s smoke false positive)
- audit 3rd (1h long-run + interval SIGUSR1): ASan 1.22 MB / 10650 alloc - runtime leak 확정
  * 새 Direct leak 2건 (3808 byte × 119 alloc × 2 path)
  * call stack: TeardownPipeline → ResetSourceOnly → ReconnectWorker
  * 원인: GStreamer rtpmanager (rtspsrc 내부) cleanup 시 g_object_unref leak
  * 크기: ~320 byte / reconn × 12 reconn/h = ~46 KB / 12h (plateau noise 안)
  * v3 의심 "매 reconn 5MB" 가설 정정 (실제 ~320 byte)

잠재 14 fix 적용 (사용자 결정: exception-escape 는 try/catch 흡수):
- code/Engine/NPU/YoloV5_Torch_Onnx_RKNN_NPU.cpp:
  * line 517: use-after-move (input.size → target_input.size)
  * line 200/204/209/210/214/222: bugprone-implicit-widening (static_cast<size_t>)
  * line 571: 동일 (data + static_cast<size_t>(i) * 3)
- code/Tracker/SORT/MgenHungarian.cpp:
  * line 175/186/235: 동일 (nOfRows * col → static_cast<size_t>(nOfRows) * col)
- code/Main/DETECTOR/src/RtspDetectorUnit.cpp:
  * line 875: misplaced-widening-cast (2 * detect_fps_limit_ → 2 * static_cast<size_t>(detect_fps_limit_))
- code/Management/manager/src/SettingData.cpp:
  * line 344/350: exception-escape (CameraSettingData ctor try/catch 흡수, CLAUDE.md "external throw 흡수" 패턴)

일부 안전 cleanup 3건 (별도 PR 의도였으나 잠재 fix 와 같이 적용):
- code/AbnormalActions/src/AbnormalActionChecker.cpp: const 제거 (performance-move-const-arg)
- code/BasicLibs/core/types/InferObject.cpp: bbox std::move 제거 (trivially-copyable)

문서:
- logs/SESSION_DFPS_B3_B4_PLATEAU_20260519.md: audit phase 추가 (1st/2nd/3rd)
- logs/AUDIT_REPORT_20260519.md: 신규 (5종 도구 결과 + 잠재 14 + 안전 152 분류 + runtime leak 분석)

build verify: OK (warning 0, error 0).

나머지 안전 70+ 는 별도 cleanup PR (refactor/audit-cleanup) — 사용자 결정.
GStreamer rtpmanager runtime leak 처리 방향 = 사용자 결정 (수용 vs 강화 vs upgrade).

+++ code/Tracker/SORT/MgenHungarian.cpp
@@ -172,7 +172,8 @@ namespace MGEN
 
 			for( col = 0; col < nOfColumns; col++ ) {
 				/* find the smallest element in the column */
-				ptrDistMatrixCur = memory_unit->dist_matrix + nOfRows * col;
+				// bugprone-implicit-widening: int*int → ptrdiff_t cast 명시 (큰 matrix 대비 방어).
+				ptrDistMatrixCur = memory_unit->dist_matrix + static_cast<size_t>( nOfRows ) * col;
 				ptrColumnEnd = ptrDistMatrixCur + nOfRows;
 
 				minValue = *ptrDistMatrixCur++;
@@ -183,7 +184,7 @@ namespace MGEN
 				}
 
 				/* subtract the smallest element from each element of the column */
-				ptrDistMatrixCur = memory_unit->dist_matrix + nOfRows * col;
+				ptrDistMatrixCur = memory_unit->dist_matrix + static_cast<size_t>( nOfRows ) * col;
 				while( ptrDistMatrixCur < ptrColumnEnd )
 					*ptrDistMatrixCur++ -= minValue;
 			}
@@ -232,7 +233,8 @@ namespace MGEN
 
 		/* cover every column containing a starred zero */
 		for( col = 0; col < nOfColumns; col++ ) {
-			starMatrixTemp = starMatrix + nOfRows * col;
+			// bugprone-implicit-widening: int*int → ptrdiff_t cast 명시.
+			starMatrixTemp = starMatrix + static_cast<size_t>( nOfRows ) * col;
 			ptrColumnEnd = starMatrixTemp + nOfRows;
 			while( starMatrixTemp < ptrColumnEnd ) {
 				if( *starMatrixTemp++ ) {
